fix: Add diagnostic header oon the design page
