+++ Composer/packages/client/src/pages/design/DesignPage.tsx
@@ -28,8 +28,7 @@ import { dialogStyle } from '../../components/Modal/dialogStyle';
 import { OpenConfirmModal } from '../../components/Modal/ConfirmDialog';
 import { ProjectTree, TreeLink } from '../../components/ProjectTree/ProjectTree';
 import { Toolbar, IToolbarItem } from '../../components/Toolbar';
-import { getFocusPath } from '../../utils/navigation';
-import { navigateTo } from '../../utils/navigation';
+import { createDiagnosticsPageUrl, getFocusPath, navigateTo } from '../../utils/navigation';
 import { getFriendlyName } from '../../utils/dialogUtil';
 import { useShell } from '../../shell';
 import plugins, { mergePluginConfigs } from '../../plugins';
@@ -61,6 +60,7 @@ import { RepairSkillModalOptionKeys } from '../../components/RepairSkillModal';
 import { useBotOperations } from '../../components/BotRuntimeController/useBotOperations';
 import { undoStatusSelectorFamily } from '../../recoilModel/selectors/undo';
 import { exportSkillModalInfoState } from '../../recoilModel/atoms/appState';
+import { DiagnosticsHeader } from '../../components/DiagnosticsHeader';
 
 import CreationModal from './creationModal';
 import { WarningMessage } from './WarningMessage';
@@ -496,6 +496,11 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
   };
 
   const toolbarItems: IToolbarItem[] = [
+    {
+      type: 'element',
+      element: <DiagnosticsHeader onClick={() => navigateTo(createDiagnosticsPageUrl(rootProjectId))} />,
+      align: 'right',
+    },
     {
       type: 'dropdown',
       text: formatMessage('Add'),
